2015.09.16. Added feature to move to screen coordinate
Previously error was raised when user double clicks on the screen.
Now if a centring position is clicked then gonio moves to this position.
If user clicks on the screen then sample is moved to clicked screen coordinate.

Double click callback is assigned to method move_to_coord
On MiniDiff this method is moveToBeam. So for compatability self.move_to_coord = self.moveToBeam
Suggestion would be to use more clear name move_to_coord